Fix custom backend for targets without atomics
